공지 게시글 상세조회

+++ catdogshelter-post-service/src/main/java/com/backoven/catdogshelter/domain/notice/query/dto/NoticeDetailDTO.java
@@ -1,4 +1,3 @@
-// src/main/java/com/backoven/catdogshelter/domain/notice/query/dto/NoticeDetailDTO.java
 package com.backoven.catdogshelter.domain.notice.query.dto;
 
 import lombok.*;
@@ -14,7 +13,7 @@ public class NoticeDetailDTO {
     private String createdAt;    // notice.created_at
     private String updatedAt;    // notice.updated_at
     private Integer ratingId;    // notice.rating_id
-    private Long likeCount;      // 공지 좋아요 수 (noticeLiked 카운트)
+    private Integer likeCount;      // 공지 좋아요 수 (noticeLiked 카운트)
 
     // 첨부파일들
     private List<NoticeFileDTO> files;
